ListView and DetailView
Created class-based views for the list of posts and a detailed view for each individual post,  PostListView doesn't follow convention for default ListView naming so the template name, context object name was changed, and post list ordered by most recently posted. PostDetailView follows the convention with the new template post_detail.html and 'object' as context. Cool thing is that it's possible to use variables in the routing, see urlpatterns for PostDetailView and its template