Email and Password Reset
Added functionality to reset password, Django makes it easy with pre-existing views in django.contrib.auth. Templates have to be made and routes have to be set, but other than that there is a view to reset the password, view to alert that an email with a link to reset password (uidb64 + token) has been sent, a view to confirm the new password to set, and a view to alert the user that the password has been reset.

Also learned how to add an App Password to sign in from apps such as this Python app,  using environment variables/decouple library to hide sensitive information, and setup the backend to generate the password reset email